Trying to make BoardEditor Bugged code, not finished
